+++ shaders/src/shadowing.fs
@@ -6,6 +6,7 @@
 #define SHADOW_SAMPLING_RUNTIME_PCF     0u
 #define SHADOW_SAMPLING_RUNTIME_VSM     1u
 #define SHADOW_SAMPLING_RUNTIME_DPCF    2u
+#define SHADOW_SAMPLING_RUNTIME_PCSS    3u
 
 //------------------------------------------------------------------------------
 // PCF Shadow Sampling
@@ -53,12 +54,33 @@ float ShadowSample_PCF(const mediump sampler2DArray shadowMap,
 // DPCF sampling
 //------------------------------------------------------------------------------
 
-const uint DPCF_SHADOW_TAP_COUNT = 12u;   // 12 max
-const vec2 poissonDisk[12] = vec2[12](
-    vec2(-0.326,-0.406), vec2(-0.840,-0.074), vec2(-0.696, 0.457), vec2(-0.203, 0.621),
-    vec2( 0.962,-0.195), vec2( 0.473,-0.480), vec2( 0.519, 0.767), vec2( 0.185,-0.893),
-    vec2( 0.507, 0.064), vec2( 0.896, 0.412), vec2(-0.322,-0.933), vec2(-0.792,-0.598)
+// Poisson disk generated with 'poisson-disk-generator' tool from
+// https://github.com/corporateshark/poisson-disk-generator by Sergey Kosarevsky
+/*const*/ mediump vec2 poissonDisk[64] = vec2[]( // don't use 'const' b/c of OSX GL compiler bug
+    vec2(0.511749, 0.547686), vec2(0.58929, 0.257224), vec2(0.165018, 0.57663), vec2(0.407692, 0.742285),
+    vec2(0.707012, 0.646523), vec2(0.31463, 0.466825), vec2(0.801257, 0.485186), vec2(0.418136, 0.146517),
+    vec2(0.579889, 0.0368284), vec2(0.79801, 0.140114), vec2(-0.0413185, 0.371455), vec2(-0.0529108, 0.627352),
+    vec2(0.0821375, 0.882071), vec2(0.17308, 0.301207), vec2(-0.120452, 0.867216), vec2(0.371096, 0.916454),
+    vec2(-0.178381, 0.146101), vec2(-0.276489, 0.550525), vec2(0.12542, 0.126643), vec2(-0.296654, 0.286879),
+    vec2(0.261744, -0.00604975), vec2(-0.213417, 0.715776), vec2(0.425684, -0.153211), vec2(-0.480054, 0.321357),
+    vec2(-0.0717878, -0.0250567), vec2(-0.328775, -0.169666), vec2(-0.394923, 0.130802), vec2(-0.553681, -0.176777),
+    vec2(-0.722615, 0.120616), vec2(-0.693065, 0.309017), vec2(0.603193, 0.791471), vec2(-0.0754941, -0.297988),
+    vec2(0.109303, -0.156472), vec2(0.260605, -0.280111), vec2(0.129731, -0.487954), vec2(-0.537315, 0.520494),
+    vec2(-0.42758, 0.800607), vec2(0.77309, -0.0728102), vec2(0.908777, 0.328356), vec2(0.985341, 0.0759158),
+    vec2(0.947536, -0.11837), vec2(-0.103315, -0.610747), vec2(0.337171, -0.584), vec2(0.210919, -0.720055),
+    vec2(0.41894, -0.36769), vec2(-0.254228, -0.49368), vec2(-0.428562, -0.404037), vec2(-0.831732, -0.189615),
+    vec2(-0.922642, 0.0888026), vec2(-0.865914, 0.427795), vec2(0.706117, -0.311662), vec2(0.545465, -0.520942),
+    vec2(-0.695738, 0.664492), vec2(0.389421, -0.899007), vec2(0.48842, -0.708054), vec2(0.760298, -0.62735),
+    vec2(-0.390788, -0.707388), vec2(-0.591046, -0.686721), vec2(-0.769903, -0.413775), vec2(-0.604457, -0.502571),
+    vec2(-0.557234, 0.00451362), vec2(0.147572, -0.924353), vec2(-0.0662488, -0.892081), vec2(0.863832, -0.407206)
 );
+
+// tap count up can go up to 64
+const uint DPCF_SHADOW_TAP_COUNT                = 12u;
+// more samples lead to better "shape" of the hardened shadow
+const uint PCSS_SHADOW_BLOCKER_SEARCH_TAP_COUNT = 16u;
+// less samples lead to noisier shadows (can be mitigated with TAA)
+const uint PCSS_SHADOW_FILTER_TAP_COUNT         = 16u;
 
 float random(const highp vec2 w) {
     const vec3 m = vec3(0.06711056, 0.00583715, 52.9829189);
@@ -95,7 +117,7 @@ mat2 getRandomRotationMatrix(highp vec2 fragCoord) {
     return R;
 }
 
-float getPenumbra(const bool DIRECTIONAL, const uint index, const highp float zLight) {
+float getPenumbraLs(const bool DIRECTIONAL, const uint index, const highp float zLight) {
     float penumbra;
     // This conditional is resolved at compile time
     if (DIRECTIONAL) {
@@ -126,7 +148,64 @@ float getPenumbraRatio(const bool DIRECTIONAL, const uint index,
         float nearOverFarMinusNear = shadowUniforms.shadows[index].nearOverFarMinusNear;
         penumbraRatio = (nearOverFarMinusNear + z_blocker) / (nearOverFarMinusNear + z_receiver) - 1.0;
     }
-    return penumbraRatio;
+    return penumbraRatio * frameUniforms.shadowPenumbraRatioScale;
+}
+
+void blockerSearchAndFilter(out float occludedCount, out float z_occSum,
+        const mediump sampler2DArray map, const highp vec2 uv, const float z_rec, const uint layer,
+        const highp vec2 filterRadii, const mat2 R, const highp vec2 dz_duv,
+        const uint tapCount) {
+    occludedCount = 0.0;
+    z_occSum = 0.0;
+    for (uint i = 0u; i < tapCount; i++) {
+        highp vec2 duv = R * (poissonDisk[i] * filterRadii);
+        float z_occ = textureLod(map, vec3(uv + duv, layer), 0.0).r;
+
+        // note: z_occ and z_rec are not necessarily linear here, comparing them is always okay for
+        // the regular PCF, but the "distance" is meaningless unless they are actually linear
+        // (e.g.: for the directional light).
+        // Either way, if we assume that all the samples are close to each other we can take their
+        // average regardless, and the average depth value of the occluders
+        // becomes: z_occSum / occludedCount.
+
+        // receiver plane depth bias
+        float z_bias = dot(dz_duv, duv);
+        float dz = z_occ - z_rec; // dz>0 when blocker is between receiver and light
+        float occluded = step(z_bias, dz);
+        occludedCount += occluded;
+        z_occSum += z_occ * occluded;
+    }
+}
+
+float filterPCSS(const mediump sampler2DArray map, const highp vec2 size,
+        const highp vec2 uv, const float z_rec, const uint layer,
+        const highp vec2 filterRadii, const mat2 R, const highp vec2 dz_duv,
+        const uint tapCount) {
+
+    float occludedCount = 0.0;
+    for (uint i = 0u; i < tapCount; i++) {
+        highp vec2 duv = R * (poissonDisk[i] * filterRadii);
+
+        // sample the shadow map with a 2x2 PCF, this helps a lot in low resolution areas
+        vec4 d;
+        highp vec2 st = (uv + duv) * size - 0.5;
+        highp vec2 grad = fract(st);
+#if defined(FILAMENT_HAS_FEATURE_TEXTURE_GATHER)
+        d = textureGather(map, vec3(uv + duv, layer), 0); // 01, 11, 10, 00
+#else
+        d[0] = texelFetchOffset(map, ivec3(st, layer), 0, ivec2(0, 1)).r;
+        d[1] = texelFetchOffset(map, ivec3(st, layer), 0, ivec2(1, 1)).r;
+        d[2] = texelFetchOffset(map, ivec3(st, layer), 0, ivec2(1, 0)).r;
+        d[3] = texelFetchOffset(map, ivec3(st, layer), 0, ivec2(0, 0)).r;
+#endif
+
+        // receiver plane depth bias
+        float z_bias = dot(dz_duv, duv);
+        vec4 dz = d - vec4(z_rec); // dz>0 when blocker is between receiver and light
+        vec4 pcf = step(z_bias, dz);
+        occludedCount += mix(mix(pcf.w, pcf.z, grad.x), mix(pcf.x, pcf.y, grad.x), grad.y);
+    }
+    return occludedCount * (1.0 / float(tapCount));
 }
 
 /*
@@ -143,31 +222,17 @@ float ShadowSample_DPCF(const bool DIRECTIONAL,
     // large kernel.
     highp vec2 dz_duv = computeReceiverPlaneDepthBias(position);
 
-    float penumbra = getPenumbra(DIRECTIONAL, index, zLight);
+    float penumbra = getPenumbraLs(DIRECTIONAL, index, zLight);
 
     // rotate the poisson disk randomly
     mat2 R = getRandomRotationMatrix(gl_FragCoord.xy);
 
     float occludedCount = 0.0;
     float z_occSum = 0.0;
-    for (uint i = 0u; i < DPCF_SHADOW_TAP_COUNT; i++) {
-        highp vec2 duv = R * poissonDisk[i] * (texelSize * penumbra);
-        float z_occ = textureLod(map, vec3(position.xy + duv, layer), 0.0).r;
 
-        // note: z_occ and z_rec are not necessarily linear here, comparing them is always okay for
-        // the regular PCF, but the "distance" is meaningless unless they are actually linear
-        // (e.g.: for the directional light).
-        // Either way, if we assume that all the samples are close to each other we can take their
-        // average regardless, and the average depth value of the occluders
-        // becomes: z_occSum / occludedCount.
-
-        // receiver plane depth bias
-        float z_bias = dot(dz_duv, duv);
-        float dz = z_occ - position.z; // dz>0 when blocker is between receiver and light
-        float occluded = step(z_bias, dz);
-        occludedCount += occluded;
-        z_occSum += z_occ * occluded;
-    }
+    blockerSearchAndFilter(occludedCount, z_occSum,
+            map, position.xy, position.z, layer, texelSize * penumbra, R, dz_duv,
+            DPCF_SHADOW_TAP_COUNT);
 
     // early exit if there is no occluders at all, also avoids a divide-by-zero below.
     if (z_occSum == 0.0) {
@@ -189,6 +254,43 @@ float ShadowSample_DPCF(const bool DIRECTIONAL,
 
     // now we just need to lerp between hardened PCF and regular PCF based on alpha
     percentageOccluded = mix(hardenedKernel(percentageOccluded), percentageOccluded, penumbraRatio);
+    return 1.0 - percentageOccluded;
+}
+
+float ShadowSample_PCSS(const bool DIRECTIONAL,
+        const mediump sampler2DArray map, const uint layer, const uint index,
+        const highp vec4 shadowPosition, const highp float zLight) {
+    highp vec2 size = vec2(textureSize(map, 0));
+    highp vec2 texelSize = vec2(1.0) / size;
+    highp vec3 position = shadowPosition.xyz * (1.0 / shadowPosition.w);
+
+    // We need to use the shadow receiver plane depth bias to combat shadow acne due to the
+    // large kernel.
+    highp vec2 dz_duv = computeReceiverPlaneDepthBias(position);
+
+    float penumbra = getPenumbraLs(DIRECTIONAL, index, zLight);
+
+    // rotate the poisson disk randomly
+    mat2 R = getRandomRotationMatrix(gl_FragCoord.xy);
+
+    float occludedCount = 0.0;
+    float z_occSum = 0.0;
+
+    blockerSearchAndFilter(occludedCount, z_occSum,
+            map, position.xy, position.z, layer, texelSize * penumbra, R, dz_duv,
+            PCSS_SHADOW_BLOCKER_SEARCH_TAP_COUNT);
+
+    // early exit if there is no occluders at all, also avoids a divide-by-zero below.
+    if (z_occSum == 0.0) {
+        return 1.0;
+    }
+
+    float penumbraRatio = getPenumbraRatio(DIRECTIONAL, index, position.z, z_occSum / occludedCount);
+
+    float percentageOccluded = filterPCSS(map, size, position.xy, position.z, layer,
+            texelSize * (penumbra * penumbraRatio),
+            R, dz_duv, PCSS_SHADOW_FILTER_TAP_COUNT);
+
     return 1.0 - percentageOccluded;
 }
 
@@ -356,7 +458,7 @@ float shadow(const bool DIRECTIONAL,
     return ShadowSample_PCF(shadowMap, layer, shadowPosition);
 }
 
-// VSM or DPCF sampling
+// Shadow requiring a sampler2D sampler (VSM, DPCF and PCSS)
 float shadow(const bool DIRECTIONAL,
         const mediump sampler2DArray shadowMap,
         const uint layer, const uint index, const uint cascade) {
@@ -382,6 +484,10 @@ float shadow(const bool DIRECTIONAL,
 
     if (frameUniforms.shadowSamplingType == SHADOW_SAMPLING_RUNTIME_DPCF) {
         return ShadowSample_DPCF(DIRECTIONAL, shadowMap, layer, index, shadowPosition, zLight);
+    }
+
+    if (frameUniforms.shadowSamplingType == SHADOW_SAMPLING_RUNTIME_PCSS) {
+        return ShadowSample_PCSS(DIRECTIONAL, shadowMap, layer, index, shadowPosition, zLight);
     }
 
     if (frameUniforms.shadowSamplingType == SHADOW_SAMPLING_RUNTIME_PCF) {