+++ RELEASE_NOTES.md
@@ -3,7 +3,14 @@
 This file contains one line summaries of commits that are worthy of mentioning in release notes.
 A new header is inserted each time a *tag* is created.
 
-## v1.15.2 (currently main branch)
+## v1.15.3 (currently main branch)
+
+## v1.15.2
+
+- engine: add support for PCSS (Percentage Closer Soft Shadows).
+- engine: fix spotlight shadow bias.
+- samples: avoid leaking IBLs in Android sample-gltf-viewer.
+- libs: fix `libibl` on mobile.
 
 ## v1.15.1